fix: address review feedback on on-demand model loading
- Log the full traceback (not just the exception string) when an
  on-demand fetch fails and the job falls back to the primary model.
- Replace the permanent failed-fetch cache with a 300s cooldown so
  transient network errors become retryable without a server restart,
  while repeated requests inside the cooldown still fail fast.
- Split OnDemandModelLoadTests into job_model_selection_on_demand_test.py
  so both test modules sit within the 200-line module cap.
- Add coverage for non-numeric ACESTEP_QUEUE_WORKERS values (feature
  disables instead of raising) and for cooldown expiry.
